requirements.txt
pandas>=1.3.0
numpy>=1.21.0
requests>=2.25.0
yfinance>=0.1.70
python-dotenv>=0.20.0
streamlit>=1.10.0
scikit-learn>=1.0.0
joblib>=1.1.0
pandas-ta>=0.3.14b0
